refactor(client): clean up deep imports
