+++ packages/client/src/components/Group/components/Activities/activities/IframeToolbar/ReadOnlyIframeToolbar.tsx
@@ -3,10 +3,10 @@ import CloseIcon from '@material-ui/icons/Close';
 import FullscreenIcon from '@material-ui/icons/Fullscreen';
 import FullscreenExitIcon from '@material-ui/icons/FullscreenExit';
 import RefreshIcon from '@material-ui/icons/Refresh';
+import { backgroundColor } from 'lib/theme';
 import React from 'react';
 import { useRecoilState, useRecoilValue } from 'recoil';
 
-import { backgroundColor } from '../../../../../../lib/theme';
 import { pinnedStreamIdState } from '../../../../lib/GroupState';
 import { isAdminState } from '../../../../lib/useIsAdmin/useIsAdmin';
 import ActivityContextMenu from '../../ActivityContextMenu/ActivityContextMenu';
