+++ packages/client/src/components/Group/components/BasicView/components/TiledVideoLayout/TiledVideoLayout.tsx
@@ -1,8 +1,8 @@
 import { Box } from '@material-ui/core';
+import { UserDocumentData } from 'lib/hooks/useUser';
 import React from 'react';
 import useResizeObserver from 'use-resize-observer';
 
-import { UserDocumentData } from '../../../../../../lib/hooks/useUser';
 import { User } from '../../../../lib/useSocketHandler';
 import mapActivityIdToComponent from '../../../Activities/lib/mapActivityIdToComponent';
 import { ActivityId } from '../../../Activities/lib/useActivities';
