+++ packages/client/src/components/Group/components/Sidebar/components/OverflowMenu/OverflowMenu.tsx
@@ -4,10 +4,10 @@ import MoreVertIcon from '@material-ui/icons/MoreVert';
 import ScreenShareIcon from '@material-ui/icons/ScreenShare';
 import SettingsIcon from '@material-ui/icons/Settings';
 import StopScreenShareIcon from '@material-ui/icons/StopScreenShare';
+import { FEEDBACK_FORM_URL } from 'lib/constants';
 import React from 'react';
 import { useSetRecoilState } from 'recoil';
 
-import { FEEDBACK_FORM_URL } from '../../../../../../constants';
 import { mediaSettingsModalIsOpenState } from '../../../MediaSettingsModal/MediaSettingsModal';
 import * as S from './OverflowMenu.styles';
