+++ packages/client/src/components/Home/lib/useUpdateGroups.ts
@@ -1,6 +1,6 @@
+import { GroupType } from 'lib/types';
 import { useContext } from 'react';
 
-import { GroupType } from '../../../lib/types';
 import { FirebaseContext } from '../../Firebase';
 
 export default function(options?: { client?: typeof firebase }) {
