+++ packages/client/src/index.tsx
@@ -1,11 +1,11 @@
 import './index.css';
 
 import * as Sentry from '@sentry/browser';
+import { appConfig } from 'lib/constants';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
 import App from './App';
-import { appConfig } from './constants';
 import * as serviceWorker from './serviceWorker';
 
 Sentry.init({
